Changes to fix to positioning problem
When the guard is in ragdoll mode it is possible to push the ragdoll out of position inside the containing parent object and so the position in the navmesh and the position of the character are offset. Added a reset script that records the local relationship of the ragdoll to the parent and can reset them.